Fix Import/Export functionality Change importing and exporting from Base64String to filelocation Add Internet Connection check around download from URL Add Popup for BookCoverURL

+++ BookCollector/ViewModels/Groupings/CollectionsViewModel.cs
@@ -169,8 +169,14 @@ namespace BookCollector.ViewModels.Groupings
                 {
                     SetIsBusyTrue();
 
-                    // Unit test data
-                    TestData.DeleteCollection(selected);
+                    if (TestData.UseTestData)
+                    {
+                        TestData.DeleteCollection(selected);
+                    }
+                    else
+                    {
+
+                    }
 
                     await ConfirmDelete(selected.CollectionName);
